add security redirect + post datetime

+++ src/Controller/PostController.php
@@ -7,6 +7,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Form\PostType;
+use DateTime;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -37,6 +38,7 @@ class PostController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $post->setUser($this->getUser());
+            $post->setPublishedAt(new \DateTime());
             $em = $doctrine->getManager();
             $em->persist($post);
             $em->flush();
@@ -51,6 +53,11 @@ class PostController extends AbstractController
     public function delete(Post $post, ManagerRegistry $doctrine): Response
     {
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        if ($this->getUser() !== $post->getUser()) {
+            $this->addFlash('error', 'Vous ne pouvez pas supprimer une publication qui ne vous appartient pas.');
+
+            return $this->redirectToRoute('home');
+        }
         $em = $doctrine->getManager();
         $em->remove($post);
         $em->flush();
@@ -62,6 +69,10 @@ class PostController extends AbstractController
     public function update(Request $request, Post $post, ManagerRegistry $doctrine): Response
     {
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        if ($this->getUser() !== $post->getUser()) {
+            $this->addFlash('error', 'Vous ne pouvez pas dupliquer une publication qui ne vous appartient pas.');
+            return $this->redirectToRoute('home');
+        }
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -79,8 +90,11 @@ class PostController extends AbstractController
     {
 
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        if ($this->getUser() !== $post->getUser()) {
+            $this->addFlash('error', 'Vous ne pouvez pas modifier une publication qui ne vous appartient pas.');
+            return $this->redirectToRoute('home');
+        }
         $copyPost = clone ($post);
-
         $em = $doctrine->getManager();
         $em->persist($copyPost);
         $em->flush();
